Use expand_number(3) from libutil instead of home-grown function to parse human-friendly power-of-two numbers (i.e. 2k, 5M etc).
Suggested by:	many
MFC after:	1 week


git-svn-id: svn+ssh://svn.freebsd.org/base/head@204654 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sbin/newfs/newfs.h
@@ -86,19 +86,19 @@ extern int	Jflag;		/* enable gjournal for file system */
 extern int	lflag;		/* enable multilabel MAC for file system */
 extern int	nflag;		/* do not create .snap directory */
 extern intmax_t	fssize;		/* file system size */
-extern int	sectorsize;	/* bytes/sector */
+extern int64_t	sectorsize;	/* bytes/sector */
 extern int	realsectorsize;	/* bytes/sector in hardware*/
-extern int	fsize;		/* fragment size */
-extern int	bsize;		/* block size */
-extern int	maxbsize;	/* maximum clustering */
-extern int	maxblkspercg;	/* maximum blocks per cylinder group */
+extern int64_t	fsize;		/* fragment size */
+extern int64_t	bsize;		/* block size */
+extern int64_t	maxbsize;	/* maximum clustering */
+extern int64_t	maxblkspercg;	/* maximum blocks per cylinder group */
 extern int	minfree;	/* free space threshold */
 extern int	opt;		/* optimization preference (space or time) */
-extern int	density;	/* number of bytes per inode */
-extern int	maxcontig;	/* max contiguous blocks to allocate */
-extern int	maxbpg;		/* maximum blocks per file in a cyl group */
-extern int	avgfilesize;	/* expected average file size */
-extern int	avgfilesperdir;	/* expected number of files per directory */
+extern int64_t	density;	/* number of bytes per inode */
+extern int64_t	maxcontig;	/* max contiguous blocks to allocate */
+extern int64_t	maxbpg;		/* maximum blocks per file in a cyl group */
+extern int64_t	avgfilesize;	/* expected average file size */
+extern int64_t	avgfilesperdir;	/* expected number of files per directory */
 extern u_char	*volumelabel;	/* volume label for filesystem */
 extern struct uufsd disk;	/* libufs disk structure */
